+++ .github/workflows/secrets.yaml
@@ -5,7 +5,7 @@ on:
   push:
 
     branches:
-      - secrets-temp
+      - streamlit-app
 
 jobs:
