+++ twitterbot.py
@@ -4,21 +4,11 @@ import os
 import streamlit as st 
 import re
 from sentiment_detection import mood
+from utils import get_tweet, get_twitter_api
 #credentials = yaml.load(open('TestbaseRadhika/.github/workflows/secrets.yaml'))
 auth = tweepy.OAuthHandler( os.environ['CONSUMER_KEY'] , os.environ['CONSUMER_SECRET'] )
 auth.set_access_token( os.environ['ACCESS_KEY'] , os.environ['ACCESS_SECRET'] )
 
-def get_twitter_api(auth):
-    twitter_API = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True)
-    return twitter_API
-
-def get_tweet(tweet_url, twitter_API):
-    try : 
-        tweet_id = re.findall("(?<=status\/)(.*)(?=\?s)", tweet_url)
-        tweet = twitter_API.get_status(tweet_id[0]).text
-    except:
-        raise ValueError("Invalid tweet, please try again")
-    return tweet
     
 st.title("HinglishBot: Sentiment Analysis Tool")
 st.image(
